Extract shared logic for type: "..." plugins
