+++ packages/babel-helper-import-to-platform-api/package.json
@@ -15,7 +15,8 @@
   "main": "./lib/index.js",
   "dependencies": {
     "@babel/helper-compilation-targets": "workspace:^",
-    "@babel/helper-module-imports": "workspace:^"
+    "@babel/helper-module-imports": "workspace:^",
+    "@babel/helper-plugin-utils": "workspace:^"
   },
   "peerDependencies": {
     "@babel/core": "^8.0.0"
