+++ packages/babel-plugin-proposal-bytes-modules/test/fixtures/errors/import-declaration-with-source-phase/options.json
@@ -1,5 +1,5 @@
 {
   "sourceType": "module",
   "plugins": ["proposal-bytes-modules", "syntax-import-source"],
-  "throws": "Bytes modules do not support phase modifiers."
+  "throws": "bytes modules do not support phase modifiers."
 }
