+++ packages/babel-plugin-proposal-text-modules/test/fixtures/errors/import-declaration-with-defer-phase/options.json
@@ -1,5 +1,5 @@
 {
   "sourceType": "module",
   "plugins": ["proposal-text-modules", "syntax-import-defer"],
-  "throws": "Text modules do not support phase modifiers."
+  "throws": "text modules do not support phase modifiers."
 }
