+++ packages/babel-plugin-transform-json-modules/test/fixtures/errors/import-declaration-with-source-phase/options.json
@@ -1,5 +1,5 @@
 {
   "sourceType": "module",
   "plugins": ["transform-json-modules", "syntax-import-source"],
-  "throws": "JSON modules do not support phase modifiers."
+  "throws": "json modules do not support phase modifiers."
 }
